added access filters

+++ eComH24S1/app/views/publications/view.php
@@ -49,17 +49,19 @@
                                     <p>
                                         <?php echo $comment['timestamp']; ?>
                                     </p>
-                                    <?php
-                                    $profile = new \app\models\Profile();
-                                    $commentProfile = $profile->getForUser($_SESSION['user_id']);
-                                    if ($comment['profile_id'] == $commentProfile->profile_id): ?>
-                                        <form action="/Publications/editComment/<?php echo $comment['publication_comment_id']; ?>"
-                                            method="POST">
-                                            <textarea name="new_comment_text" rows="4" cols="50"
-                                                required><?php echo $comment['comment_text']; ?></textarea><br><br>
-                                            <input type="submit" value="Save Changes">
-                                        </form>
-                                        <a href="/Publications/deleteComment/<?php echo $comment['publication_comment_id']; ?>">Delete</a>
+                                    <?php if (isset ($_SESSION['user_id'])): ?>
+                                        <?php
+                                        $profile = new \app\models\Profile();
+                                        $commentProfile = $profile->getForUser($_SESSION['user_id']);
+                                        if ($commentProfile && $comment['profile_id'] == $commentProfile->profile_id): ?>
+                                            <form action="/Publications/editComment/<?php echo $comment['publication_comment_id']; ?>"
+                                                method="POST">
+                                                <textarea name="new_comment_text" rows="4" cols="50"
+                                                    required><?php echo $comment['comment_text']; ?></textarea><br><br>
+                                                <input type="submit" value="Save Changes">
+                                            </form>
+                                            <a href="/Publications/deleteComment/<?php echo $comment['publication_comment_id']; ?>">Delete</a>
+                                        <?php endif; ?>
                                     <?php endif; ?>
                                 </div>
                             <?php endforeach; ?>
